experiment 2: resources and resource type files
resources and resource types in the pipeline will now be looked up under
resources/ and resource-types/, respectively.

this is basically what was discussed in the following issue + comment:

  https://github.com/concourse/concourse/issues/1994#issuecomment-472145314

there are many advantages to this. for one thing, it's easier to
navigate files in an editor than navigate through hundreds of lines of a
single YAML file.

another advantage is that this may be a good way to share resource
definitions across many pipelines in a project. i haven't yet decided
where pipelines end up in this restructuring and how they're configured
- i.e. whether it's a single pipeline per repo at the top
(pipeline.yml/concourse.yml/whatever) or whether a single repo can have
many pipelines, all auto-configured. this could be one reason to go for
the latter.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>