Revert "experiment 4: multiple projects in one repo"
i actually don't like this for one reason: the prs project has no need
for all of the resources that the other project does, and by not
configuring a credential manager it won't even be able to 'check' for
most of them.

separate projects should explicitly provide their own resources and
tasks instead. we should still allow multiple projects in one repo, but
it should probably be specified via an entire sub-directory, not just a
filename.

This reverts commit 0539ee448ad408588d48d62a768b33a41d2a29e8.